Added some rest and front scripts

+++ data/user.py
@@ -38,7 +38,7 @@ class User:
 
     def get_posts(self, offset=0):
         posts = get_db(f"""SELECT id FROM posts 
-                           WHERE replied IS NULL AND author<>"{self.username}" 
+                           WHERE replied IS NULL AND author="{self.username}" 
                            ORDER BY date DESC LIMIT {news_piece} OFFSET {offset}""")
         return [p.Post(post[0]) for post in posts]
 
@@ -67,9 +67,18 @@ class User:
         followings = get_db(f'SELECT whom FROM followships WHERE who="{self.username}"')
         return [User(user[0]) for user in followings]
 
+    def follow(self, user):
+        if user in [u.username for u in self.get_following()]:
+            return True
+        result = execute_db(f"""INSERT INTO followships(who, whom) 
+                                VALUES ("{self.username}", "{user}" """)
+        return result
+
     def unfollow(self, user):
+        if user not in [u.username for u in self.get_following()]:
+            return True
         result = execute_db(f"""DELETE FROM followships 
-                                WHERE who={self.username} AND whom={user.username}""")
+                                WHERE who={self.username} AND whom={user}""")
         return result
 
     def delete(self):
